requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
greenlet==2.0.0
html5lib==1.1
importlib-metadata==4.8.3
iniconfig==1.1.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
tabulate==0.8.10
tomli==1.2.3
typing_extensions==4.1.1
webencodings==0.5.1
zipp==3.6.0
